Fixes swapchain validation error
Was getting `VUID-VkPresentInfoKHR-pImageIndices-01296` on swapchain
images that were successfully acquired and had the correct image layout.
Turns out the error was spurious and seems to be related to [this
issue][1].

The fix is to not call `get_swapchain_images()` lazily with a `RefCell`
and instead call it immediately after creating the swapchain and store
the images for the entire lifetime of the swapchain.

[1]: https://github.com/KhronosGroup/Vulkan-ValidationLayers/issues/6068